requirements.txt
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
archspec==0.2.1
attrs==23.2.0
blinker==1.7.0
boltons==23.0.0
Brotli==1.0.9
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
conda==23.11.0
conda-content-trust==0.2.0
conda-libmamba-solver==23.12.0
conda-package-handling==2.2.0
conda_package_streaming==0.9.0
cryptography==41.0.7
distro==1.8.0
fastapi==0.110.0
gitdb==4.0.11
GitPython==3.1.42
idna==3.4
imbalanced-learn==0.12.0
imblearn==0.0
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.32
jsonpointer==2.1
jsonschema==4.21.1
jsonschema-specifications 2023.12.1
libmambapy==1.5.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
menuinst==2.0.1
numpy==1.26.4
packaging==23.1
pandas==2.2.1
pillow==10.2.0
pip==23.3.1
platformdirs==3.10.0
pluggy==1.0.0
protobuf==4.25.3
pyarrow==15.0.1
pycosat==0.6.6
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
ruamel.yaml==0.17.21
scikit-learn==1.4.1.post1
scipy==1.12.0
setuptools==68.2.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.36.3
streamlit==1.32.0
tenacity==8.2.3
threadpoolctl==3.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.65.0
truststore==0.8.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.18
watchdog==4.0.0
wheel==0.41.2
xgboost==2.0.3
zstandard==0.19.0